- Fix an issue when a new prefab instance is created, even though it is being called from the same prefab, generates a new unique alias, therefore generating a new hash, causing it to always generate a new unique ID even though they are the same prefabs. The cause is that when creating a new   `PrefabDocument`, the instance is created with a unique alias, even though no alias is needed. This causes the instances of the same prefab to end up with unique IDs. The fix is to supply a 'dummyalias' instead to invoke the constructor for a prefab instance that does not generate a new instance alias for the prefab. - Fix internal implementation of `htonll` on platforms that don't support it (Linux, Android) to match the output of `htonll` on platforms that do support it (Windows) - Fix a network serialization issues related to signed integer overflow and order of operations. - Added unit test for type serialization for fixed expected serialization to ensure that serialization is consistent across all supported platforms/architectures

+++ Code/Framework/AzNetworking/AzNetworking/Serialization/NetworkInputSerializer.cpp
@@ -123,20 +123,20 @@ namespace AzNetworking
     {
         m_serializerValid &= (inputValue >= minValue);
         m_serializerValid &= (inputValue <= maxValue);
-        const uint64_t valueRange = static_cast<uint64_t>(maxValue - minValue);
+        const uint64_t valueRange = static_cast<uint64_t>(maxValue) - static_cast<uint64_t>(minValue);
         if (valueRange <= AZStd::numeric_limits<uint8_t>::max())
         {
-            return SerializeBoundedValueHelper<uint8_t>(static_cast<uint8_t>(inputValue - minValue));
+            return SerializeBoundedValueHelper<uint8_t>(static_cast<uint8_t>(inputValue) - static_cast<uint8_t>(minValue));
         }
         else if (valueRange <= AZStd::numeric_limits<uint16_t>::max())
         {
-            return SerializeBoundedValueHelper<uint16_t>(static_cast<uint16_t>(inputValue - minValue));
+            return SerializeBoundedValueHelper<uint16_t>(static_cast<uint16_t>(inputValue) - static_cast<uint16_t>(minValue));
         }
         else if (valueRange <= AZStd::numeric_limits<uint32_t>::max())
         {
-            return SerializeBoundedValueHelper<uint32_t>(static_cast<uint32_t>(inputValue - minValue));
+            return SerializeBoundedValueHelper<uint32_t>(static_cast<uint32_t>(inputValue) - static_cast<uint32_t>(minValue));
         }
-        return SerializeBoundedValueHelper<uint64_t>(static_cast<uint64_t>(inputValue - minValue));
+        return SerializeBoundedValueHelper<uint64_t>(static_cast<uint64_t>(inputValue) - static_cast<uint64_t>(minValue));
     }
 
     inline uint8_t HostToNetwork(uint8_t value)
